Moved the model classes to one directory.
