+++ app/src/androidTest/java/com/codemate/koffeemate/data/local/CoffeeEventRepositoryTest.kt
@@ -16,8 +16,8 @@
 
 package com.codemate.koffeemate.data.local
 
-import com.codemate.koffeemate.data.local.models.CoffeeBrewingEvent
-import com.codemate.koffeemate.data.network.models.User
+import com.codemate.koffeemate.data.models.CoffeeBrewingEvent
+import com.codemate.koffeemate.data.models.User
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import org.hamcrest.core.IsEqual.equalTo
