+++ app/src/androidTest/java/com/codemate/koffeemate/data/local/UserRepositoryTest.kt
@@ -16,7 +16,7 @@
 
 package com.codemate.koffeemate.data.local
 
-import com.codemate.koffeemate.data.network.models.User
+import com.codemate.koffeemate.data.models.User
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import org.hamcrest.core.IsEqual.equalTo
