+++ app/src/main/java/com/codemate/koffeemate/ui/main/PostAccidentUseCase.kt
@@ -20,8 +20,8 @@ import android.graphics.Bitmap
 import com.codemate.koffeemate.common.AwardBadgeCreator
 import com.codemate.koffeemate.data.local.CoffeeEventRepository
 import com.codemate.koffeemate.data.local.CoffeePreferences
+import com.codemate.koffeemate.data.models.User
 import com.codemate.koffeemate.data.network.SlackApi
-import com.codemate.koffeemate.data.network.models.User
 import com.codemate.koffeemate.extensions.toRequestBody
 import okhttp3.MediaType
 import okhttp3.MultipartBody
