+++ app/src/main/java/com/codemate/koffeemate/ui/userselector/UserSelectorAdapter.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import com.bumptech.glide.Glide
 import com.codemate.koffeemate.R
-import com.codemate.koffeemate.data.network.models.User
+import com.codemate.koffeemate.data.models.User
 import kotlinx.android.synthetic.main.recycler_item_user.view.*
 import org.jetbrains.anko.onClick
 import java.util.*
