+++ app/src/main/java/com/codemate/koffeemate/ui/userselector/UserSelectorPresenter.kt
@@ -1,6 +1,6 @@
 package com.codemate.koffeemate.ui.userselector
 
-import com.codemate.koffeemate.data.network.models.User
+import com.codemate.koffeemate.data.models.User
 import com.codemate.koffeemate.ui.base.BasePresenter
 import rx.Subscriber
 import javax.inject.Inject
