+++ app/src/test/java/com/codemate/koffeemate/ui/userselector/LoadUsersUseCaseTest.kt
@@ -18,9 +18,9 @@ package com.codemate.koffeemate.ui.userselector
 
 import com.codemate.koffeemate.BuildConfig
 import com.codemate.koffeemate.data.local.UserRepository
+import com.codemate.koffeemate.data.models.User
 import com.codemate.koffeemate.data.network.SlackApi
 import com.codemate.koffeemate.data.network.SlackService
-import com.codemate.koffeemate.data.network.models.User
 import com.codemate.koffeemate.testutils.getResourceFile
 import com.nhaarman.mockito_kotlin.verify
 import com.nhaarman.mockito_kotlin.whenever
